+++ lib/prawn/emoji/drawable.rb
@@ -5,9 +5,16 @@ require_relative 'drawer'
 module Prawn
   module Emoji
     module Drawable
+      # == Additional Options
+      # <tt>:emoji</tt>:: <tt>boolean</tt>. Whether or not to draw an emoji [true]
       def draw_text!(text, options)
-        text_without_emoji = emoji_drawer.draw(text.to_s, options)
-        super text_without_emoji, options
+        draw_emoji = options.fetch(:emoji, true)
+
+        if draw_emoji && text.encoding == ::Encoding::UTF_8 && Emoji.regex.match?(text)
+          emoji_drawer.draw(text.to_s, options)
+        else
+          super(text, options)
+        end
       end
 
       private