+++ lib/prawn/emoji/drawer.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require_relative 'substitution'
 require_relative 'index'
 require_relative 'image'
 require_relative 'text'
@@ -14,41 +13,40 @@ module Prawn
       end
 
       def draw(text, text_options)
-        return text unless text.encoding == ::Encoding::UTF_8
-        return text unless Emoji.regex.match?(text)
+        cursor_x, cursor_y = text_options[:at]
 
-        result = []
-        target = Emoji::Text.new(text)
+        emoji_text = Emoji::Text.new(text, document.font_size)
 
-        while target.contains_emoji? do
-          if emoji_index.include?(target.emoji.codepoint)
-            draw_emoji(
-              target,
-              text_options: text_options,
-              base_text: result.join
-            )
-            result << target.left + Emoji::Substitution.new(@document).to_s
+        while emoji_text.contains_emoji? do
+          if emoji_index.include?(emoji_text.emoji.codepoint)
+            cursor_x += draw_text(emoji_text.left, at: [cursor_x, cursor_y], text_options: text_options)
+            cursor_x += draw_emoji(emoji_text.emoji, at: [cursor_x, cursor_y])
           else
-            result << target.left_with_emoji
+            cursor_x += draw_text(emoji_text.left_with_emoji, at: [cursor_x, cursor_y], text_options: text_options)
           end
-
-          target = Emoji::Text.new(target.remaining)
+          emoji_text = Emoji::Text.new(emoji_text.remaining, document.font_size)
         end
 
-        result.join + target.to_s
+        draw_text!(emoji_text.to_s, at: [cursor_x, cursor_y], text_options: text_options)
       end
 
       private
 
-      attr_reader :emoji_index
+      attr_reader :emoji_index, :document
 
-      def draw_emoji(text, text_options:, base_text:)
-        image = Emoji::Image.new(text.emoji, @document.font_size)
+      def draw_text!(text, at:, text_options:)
+        document.draw_text!(text, text_options.merge(emoji: false, at: at))
+      end
 
-        base_x, base_y = text_options[:at]
+      def draw_text(text, at:, text_options:)
+        draw_text!(text, at: at, text_options: text_options)
+        document.width_of(text, text_options)
+      end
 
-        x = image.adjust_x(base_x + @document.width_of(base_text + text.left, text_options))
-        y = image.adjust_y(base_y)
+      def draw_emoji(emoji_char, at:)
+        emoji_image = Emoji::Image.new(emoji_char)
+
+        x, y = at
 
         # Prawn 2.2 does not close the image file when Pathname is passed to Document#image method.
         #
@@ -57,9 +55,11 @@ module Prawn
         #
         #   @document.image(image_file.path, at: [x, y], width: image.width)
         #
-        File.open(image.path, 'rb') do |image_file|
-          @document.image(image_file, at: [x, y], width: image.width)
+        File.open(emoji_image.path, 'rb') do |image_file|
+          @document.image(image_file, at: [x, y + emoji_char.height], width: emoji_char.width)
         end
+
+        emoji_char.width + document.character_spacing
       end
     end
   end